feat: split into multiple packages
BREAKING CHANGE: `@faergeek/eslint-config` now only exports what
previously was the `base` export as it's `default` export. For other
exports a `@faergeek/eslint-config-<export>` package needs to be
installed.
