+++ package.json
@@ -1,40 +1,17 @@
 {
+  "private": true,
   "type": "module",
-  "name": "@faergeek/eslint-config",
-  "version": "7.1.6",
-  "exports": "./index.js",
-  "files": [
-    "index.js"
-  ],
-  "author": "Sergei Slipchenko <faergeek@gmail.com>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/faergeek/eslint-config.git"
-  },
   "scripts": {
     "lint": "eslint --max-warnings 0",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "generate-changelog": "git-cliff --bump -o CHANGELOG.md && prettier --write CHANGELOG.md",
-    "version": "pnpm run generate-changelog && git add CHANGELOG.md"
+    "generate-changelog": "git-cliff --bump -o CHANGELOG.md && prettier --write CHANGELOG.md"
   },
   "packageManager": "pnpm@10.15.1",
-  "dependencies": {
-    "@eslint/js": "^9.19.0",
-    "@vitest/eslint-plugin": "^1.0.1",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-simple-import-sort": "^12.0.0",
-    "globals": "^16.0.0",
-    "typescript-eslint": "^8.0.0"
-  },
-  "peerDependencies": {
-    "eslint": "^9.0.0"
-  },
   "devDependencies": {
     "@commitlint/cli": "^19.0.0",
     "@commitlint/config-conventional": "^19.0.0",
+    "@faergeek/eslint-config": "workspace:",
     "eslint": "^9.0.0",
     "git-cliff": "^2.10.0",
     "prettier": "^3.0.0"
