+++ workspace/eslint-config/package.json
@@ -1,0 +1,26 @@
+{
+  "type": "module",
+  "name": "@faergeek/eslint-config",
+  "version": "7.1.6",
+  "exports": "./index.js",
+  "files": [
+    "index.js"
+  ],
+  "author": "Sergei Slipchenko <faergeek@gmail.com>",
+  "license": "MIT",
+  "homepage": "https://github.com/faergeek/eslint-config/tree/main/workspace/eslint-config",
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/faergeek/eslint-config.git"
+  },
+  "scripts": {
+    "set-version": "npm version"
+  },
+  "dependencies": {
+    "@eslint/js": "^9.19.0",
+    "eslint-plugin-simple-import-sort": "^12.0.0"
+  },
+  "peerDependencies": {
+    "eslint": "^9.0.0"
+  }
+}
